intermediate comit restructuresd a lot
