+++ src/BuzzerScript/src/LEDController.hpp
@@ -6,23 +6,27 @@
 #include "SerialHandler.hpp"
 #include "Config.hpp"
 
+
+
 class LEDController
 {
 private:
     Adafruit_NeoPixel strip = Adafruit_NeoPixel(NUM_LEDS, LED_PIN, NEO_GRB + NEO_KHZ800);
-
+/*
     // Set
     String JsonSetHandler(JsonDocument pJson);
     String JsonSetCollor(JsonDocument pJson);
     // Get
     String JsonGetHandler(JsonDocument pJson);
     String JsonGetCollor(JsonDocument pJson);
-
-    void SetLED(int id, int R, int G, int B);
+*/
+    
 public:
     LEDController();
     ~LEDController();
     void Setup();
-    
+
+    void SetLED(int id, int R, int G, int B);
     String JsonHandlerGetSet(JsonDocument pJson);
 };
+
