+++ src/BuzzerScript/src/SerialHandler.hpp
@@ -5,6 +5,7 @@
 
 #include "Config.hpp"
 
+
 void EventSender(String type, int ID, bool oldVal, bool newVal);
 String ErrorBuilder(String pError, bool pCritical);
 String DebugBuilder(String pSorce, String pMSG, bool pCritical, String pValue);
